requirements.txt
flask
google-cloud-storage
flask_httpauth
pymongo[srv]
